+++ Assets/Scripts/Application/Installer/KeyValueStoreInstaller.cs
@@ -2,8 +2,8 @@ using System;
 using System.Collections.Generic;
 using CAFU.KeyValueStore.Application.Interface;
 using CAFU.KeyValueStore.Application.Master;
-using CAFU.KeyValueStore.Data.DataStore.Implement;
-using CAFU.KeyValueStore.Data.Repository.Implement;
+using CAFU.KeyValueStore.Data.Implement.DataStore;
+using CAFU.KeyValueStore.Data.Implement.Repository;
 using JetBrains.Annotations;
 using Zenject;
