+++ Assets/Scripts/Data/Implement/DataStore/Memory.cs
@@ -1,12 +1,12 @@
 using System;
 using System.Collections.Generic;
 using System.Threading;
-using CAFU.KeyValueStore.Data.Repository.Interface.DataStore;
+using CAFU.KeyValueStore.Data.Interface;
 using JetBrains.Annotations;
 using UniRx.Async;
 using Zenject;
 
-namespace CAFU.KeyValueStore.Data.DataStore.Implement
+namespace CAFU.KeyValueStore.Data.Implement.DataStore
 {
     [UsedImplicitly]
     internal class Memory : IAsyncGetter, IAsyncSetter, IAsyncChecker
