+++ Assets/Scripts/Data/Interface/IAsyncChecker.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using UniRx.Async;
 
-namespace CAFU.KeyValueStore.Data.Repository.Interface.DataStore
+namespace CAFU.KeyValueStore.Data.Interface
 {
     internal interface IAsyncChecker
     {
